+++ frontend/src/Book/Details/BookDetailsPageConnector.js
@@ -9,6 +9,7 @@ import NotFound from 'Components/NotFound';
 import PageContent from 'Components/Page/PageContent';
 import PageContentBody from 'Components/Page/PageContentBody';
 import translate from 'Utilities/String/translate';
+import { fetchBooks } from 'Store/Actions/bookActions';
 import BookDetailsConnector from './BookDetailsConnector';
 
 function createMapStateToProps() {
@@ -20,30 +21,21 @@ function createMapStateToProps() {
       const titleSlug = match.params.titleSlug;
       const isFetching = books.isFetching || author.isFetching;
       const isPopulated = books.isPopulated && author.isPopulated;
-
-      // if books have been fetched, make sure requested one exists
-      // otherwise don't map titleSlug to trigger not found page
-      if (!isFetching && isPopulated) {
-        const bookIndex = _.findIndex(books.items, { titleSlug });
-        if (bookIndex === -1) {
-          return {
-            isFetching,
-            isPopulated
-          };
-        }
-      }
+      const bookExists = _.some(books.items, { titleSlug });
 
       return {
         titleSlug,
         isFetching,
-        isPopulated
+        isPopulated,
+        bookExists
       };
     }
   );
 }
 
 const mapDispatchToProps = {
-  push
+  push,
+  fetchBooks
 };
 
 class BookDetailsPageConnector extends Component {
@@ -59,11 +51,21 @@ class BookDetailsPageConnector extends Component {
     this.populate();
   }
 
+  componentDidUpdate(prevProps) {
+    if (prevProps.titleSlug !== this.props.titleSlug) {
+      this.populate();
+    }
+  }
+
   //
   // Control
 
   populate = () => {
     this.setState({ hasMounted: true });
+
+    if (this.props.titleSlug) {
+      this.props.fetchBooks({ titleSlug: this.props.titleSlug, useUserLibrary: false });
+    }
   };
 
   //
@@ -73,7 +75,8 @@ class BookDetailsPageConnector extends Component {
     const {
       titleSlug,
       isFetching,
-      isPopulated
+      isPopulated,
+      bookExists
     } = this.props;
 
     if (!titleSlug) {
@@ -84,8 +87,7 @@ class BookDetailsPageConnector extends Component {
       );
     }
 
-    if ((isFetching || !this.state.hasMounted) ||
-        (!isFetching && !isPopulated)) {
+    if (isFetching || !this.state.hasMounted) {
       return (
         <PageContent title={translate('Loading')}>
           <PageContentBody>
@@ -95,7 +97,15 @@ class BookDetailsPageConnector extends Component {
       );
     }
 
-    if (!isFetching && isPopulated && this.state.hasMounted) {
+    if (!bookExists && !isFetching) {
+      return (
+        <NotFound
+          message={translate('SorryThatBookCannotBeFound')}
+        />
+      );
+    }
+
+    if (!isFetching && isPopulated && this.state.hasMounted && bookExists) {
       return (
         <BookDetailsConnector
           titleSlug={titleSlug}